refactor: change map directory
