+++ src/app/shared/components/map/map.component.html
@@ -9,8 +9,8 @@
     <!-- Lecturer's Location Marker -->
     <map-marker
       [position]="{
-        lat: lecturerLocation().latitude,
-        lng: lecturerLocation().longitude
+        lat: userLocation().latitude,
+        lng: userLocation().longitude
       }"
       [label]="{
         text: 'You',
