+++ src/app/shared/components/map/map.component.ts
@@ -1,10 +1,7 @@
 import { Component, inject, output, signal } from '@angular/core';
-import {
-  IUserLocation,
-  IStudentLocation,
-} from '../../../../interfaces/location.interface';
-import { LocationService } from '../../../../services/location/location.service';
 import { GoogleMapsModule } from '@angular/google-maps';
+import { IUserLocation } from '../../../pages/lecturer/interfaces/location.interface';
+import { LocationService } from '../../../pages/lecturer/services/location/location.service';
 @Component({
   selector: 'liaison-map',
   standalone: true,
@@ -13,9 +10,9 @@ import { GoogleMapsModule } from '@angular/google-maps';
   styleUrl: './map.component.scss',
 })
 export class MapComponent {
-  clickedMarker = output<IStudentLocation>();
+  clickedMarker = output();
 
-  lecturerLocation = signal<IUserLocation>({
+  userLocation = signal<IUserLocation>({
     latitude: 7.9465,
     longitude: -1.0232,
   });
@@ -23,7 +20,7 @@ export class MapComponent {
   center: google.maps.LatLngLiteral = { lat: 7.9465, lng: -1.0232 };
   map: google.maps.Map | undefined;
 
-  studentLocations: IStudentLocation[] = [
+  studentLocations= [
     { lat: 7.9465, lng: -1.0232, label: 'Student 1' },
     { lat: 8.9465, lng: -1.0232, label: 'Student 2' },
     { lat: 6.9465, lng: -2.0232, label: 'Student 3' },
@@ -34,7 +31,7 @@ export class MapComponent {
 
   ngOnInit(): void {
     this.locationService.getUserLocation().subscribe((location) => {
-      this.lecturerLocation.set(location);
+      this.userLocation.set(location);
     });
   }
 
@@ -43,14 +40,14 @@ export class MapComponent {
   }
 
   calculateRouteToStudent(studentLocation: google.maps.LatLngLiteral): void {
-    const lecturerLocation: google.maps.LatLngLiteral = {
-      lat: this.lecturerLocation().latitude,
-      lng: this.lecturerLocation().longitude,
+    const userLocation: google.maps.LatLngLiteral = {
+      lat: this.userLocation().latitude,
+      lng: this.userLocation().longitude,
     };
 
     if (this.map) {
       this.locationService.calculateRoute(
-        lecturerLocation,
+        userLocation,
         studentLocation,
         this.map
       );
@@ -59,7 +56,7 @@ export class MapComponent {
     }
   }
 
-  onMarkerClick(location: IStudentLocation) {
+  onMarkerClick(location: any) {
     this.clickedMarker.emit(location);
   }
 }
